revamped ClipCard audio recording
includes new AudioRecorder which records audio and manages a visualizer
and MediaRecorderWrapper which simplifies the MediaRecorder interface
and will make for easier MediaCodec swap if necessary.

TODO Add some countdown before audio recording begins, hopefully with
level feedback.

+++ lib/src/main/java/scal/io/liger/view/Util.java
@@ -37,7 +37,7 @@ public class Util {
     /**
      * @return a String of form "MM:SS" from a raw ms value
      */
-    public static String makeTimeString(int timeMs) {
+    public static String makeTimeString(long timeMs) {
         long second = (timeMs / 1000) % 60;
         long minute = (timeMs / (1000 * 60)) % 60;
